requirements.txt
pillow
python-docx
tk
dill
customtkinter
CTkListbox 
pyglet
-e .